server/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-js-asset==1.2.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
email-validator==1.1.3
idna==3.3
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
whitenoise==5.3.0
